+++ packages/coordinator/tests/services/communication/spine-communication.spec.ts
@@ -140,8 +140,8 @@ describe("Spine communication", () => {
           details: {
             coding: [
               {
-                code: "INTERACTION_NOT_SUPPORTED_BY_MTLS_CLIENT",
-                display: "Interaction not supported by mtls client",
+                code: "INTERACTION_NOT_SUPPORTED",
+                display: "Interaction not supported",
                 system: "https://fhir.nhs.uk/R4/CodeSystem/Spine-ErrorOrWarningCode",
                 version: "1"
               }
